RDKEMW-15910: Xi6 build fix

+++ classes/legacy_entos_support_patch.bbclass
@@ -8,27 +8,40 @@
 ROOTFS_POSTPROCESS_COMMAND += ' legacy_entos_support; '
 
 legacy_entos_support(){
-    if [ -f "${IMAGE_ROOTFS}${sysconfdir}/common.properties" ]; then
+    bbnote "Execute postprocess - legacy_entos_support"
+    if [ -f "${IMAGE_ROOTFS}/${sysconfdir}/common.properties" ]; then
         bbnote "common.properties file from Vendor Layer in rootfs"
-        rm -f ${IMAGE_ROOTFS}${sysconfdir}/common-generic.properties
+        rm -f ${IMAGE_ROOTFS}/${sysconfdir}/common-generic.properties
     else
-        bbnote "Installing common.properties file from Middleware Layer"
-        mv ${IMAGE_ROOTFS}/${sysconfdir}/common-generic.properties ${IMAGE_ROOTFS}/${sysconfdir}/common.properties
+        if [ -f "${IMAGE_ROOTFS}/${sysconfdir}/common-generic.properties" ]; then
+            bbnote "Installing common.properties file from Middleware Layer"
+            mv ${IMAGE_ROOTFS}/${sysconfdir}/common-generic.properties ${IMAGE_ROOTFS}/${sysconfdir}/common.properties
+        else
+            bbnote "common-generic.properties not present for this target"
+        fi
     fi
     
     if [ -f "${IMAGE_ROOTFS}/lib/rdk/imageFlasher.sh" ]; then
         bbnote "imageFlasher.sh script added from Vendor Layer in rootfs"
         rm -f ${IMAGE_ROOTFS}/lib/rdk/imageFlasher_generic.sh
     else
-        bbnote "Installing imageFlasher.sh file from Middleware Layer"
-        mv ${IMAGE_ROOTFS}/lib/rdk/imageFlasher_generic.sh ${IMAGE_ROOTFS}/lib/rdk/imageFlasher.sh
+        if [ -f "${IMAGE_ROOTFS}/lib/rdk/imageFlasher_generic.sh" ]; then
+            bbnote "Installing imageFlasher.sh file from Middleware Layer"
+            mv ${IMAGE_ROOTFS}/lib/rdk/imageFlasher_generic.sh ${IMAGE_ROOTFS}/lib/rdk/imageFlasher.sh
+        else
+            bbnote "Installing imageFlasher_generic.sh not present for this target"
+        fi
     fi
 
     if [ -f "${IMAGE_ROOTFS}/lib/rdk/init-zram.sh" ]; then
         bbnote "init-zram.sh Script added from Vendor Layer in rootfs"
         rm -f ${IMAGE_ROOTFS}/lib/rdk/init-zram_generic.sh
     else
-        bbnote "Installing init-zram.sh file from Middleware Layer"
-        mv ${IMAGE_ROOTFS}/lib/rdk/init-zram_generic.sh ${IMAGE_ROOTFS}/lib/rdk/init-zram.sh
+        if [ -f "${IMAGE_ROOTFS}/lib/rdk/init-zram_generic.sh" ]; then
+            bbnote "Installing init-zram.sh file from Middleware Layer"
+            mv ${IMAGE_ROOTFS}/lib/rdk/init-zram_generic.sh ${IMAGE_ROOTFS}/lib/rdk/init-zram.sh
+        else
+            bbnote "Installing iinit-zram_generic.sh not present for this target"
+        fi
     fi    
 }
